Abstract out access to temperature and fan control
Encapsulate accessing the /dev/sonypi device and make the method abstract, so
other methods of retrieving the temperature and setting the fan speed can be
added in the future. The main motivation is to be able to use vfand daemon
with newer Vaio laptops which provide access to the fan speed via sysfs
instead.